Rename to PortableTextEditor
